+++ netwerk/protocol/http/AltDataOutputStreamChild.cpp
@@ -66,9 +66,9 @@ void AltDataOutputStreamChild::ReleaseIPDLReference() {
 
 bool AltDataOutputStreamChild::WriteDataInChunks(
     const nsDependentCSubstring& data) {
-  const uint32_t kChunkSize = 128 * 1024;
-  uint32_t next = std::min(data.Length(), kChunkSize);
-  for (uint32_t i = 0; i < data.Length();
+  const size_t kChunkSize = 128 * 1024;
+  size_t next = std::min(data.Length(), kChunkSize);
+  for (size_t i = 0; i < data.Length();
        i = next, next = std::min(data.Length(), next + kChunkSize)) {
     nsCString chunk(Substring(data, i, kChunkSize));
     if (mIPCOpen && !SendWriteData(chunk)) {
